enemy dies + money added
enemy dies with the second bullet
when an enemy dies, adds 50 dollars to the account
to add the turret to the node, one must press it and buy it from the
store, money is reduced if a turret is bought. timer for the next enemy
to spawn. lives reduced if the enemy enters the red box. if 0 lives, die